+++ packages/core/test/test-util/test-services/openapi/test-service/extension-api.d.ts
@@ -1,0 +1,7 @@
+import { OpenApiRequestBuilder } from '../../../../../src';
+import { ExtensionApi } from './openapi/api';
+export declare const TestServiceExtensionApi: {
+    niceGetFunction: () => OpenApiRequestBuilder<ExtensionApi, "niceGetFunction">;
+    nicePostFunction: () => OpenApiRequestBuilder<ExtensionApi, "nicePostFunction">;
+};
+//# sourceMappingURL=extension-api.d.ts.map