+++ packages/core/test/test-util/test-services/openapi/test-service/openapi/api.js
@@ -25,6 +25,7 @@ var __exportStar = (this && this.__exportStar) || function(m, exports) {
 Object.defineProperty(exports, "__esModule", { value: true });
 __exportStar(require("./api/default-api"), exports);
 __exportStar(require("./api/entity-api"), exports);
+__exportStar(require("./api/extension-api"), exports);
 __exportStar(require("./api/tag-dot-api"), exports);
 __exportStar(require("./api/tag-space-api"), exports);
 __exportStar(require("./api/test-case-api"), exports);