+++ packages/generator/src/generator-cli.ts
@@ -1,7 +1,7 @@
 #!/usr/bin/env node
 
 import { resolve, dirname } from 'path';
-import { createLogger } from '@sap-cloud-sdk/util';
+import { createLogger, ErrorWithCause } from '@sap-cloud-sdk/util';
 import { readFileSync } from 'fs-extra';
 import yargs from 'yargs';
 import { generate } from './generator';
@@ -14,9 +14,12 @@ const logger = createLogger({
 
 logger.info('Parsing args...');
 
-generate(parseCmdArgs()).then(() =>
-  logger.info('Generation of services finished successfully.')
-);
+generate(parseCmdArgs())
+  .then(() => logger.info('Generation of services finished successfully.'))
+  .catch(err => {
+    logger.error(new ErrorWithCause('Generation of services failed.', err));
+    process.exit(1);
+  });
 
 export function parseCmdArgs(): GeneratorOptions {
   const command = yargs.command(