+++ packages/generator/src/imports.ts
@@ -39,8 +39,7 @@ export function externalImportDeclaration(
 }
 
 export function coreImportDeclaration(
-  namedImports: string[],
-  oDataVersion: ODataVersion
+  namedImports: string[]
 ): ImportDeclarationStructure {
   return {
     kind: StructureKind.ImportDeclaration,
@@ -95,7 +94,7 @@ export function enumTypeImportDeclarations(
 // Only supports named imports
 export function mergeImportDeclarations(
   importDeclarations: ImportDeclarationStructure[]
-) {
+): ImportDeclarationStructure[] {
   return importDeclarations
     .reduce(
       (mergedDeclarations: ImportDeclarationStructure[], importDeclaration) => {