+++ test-packages/nightly-tests/jest.config.js
@@ -3,5 +3,6 @@ const commonConfig = require('../../test-resources/jest.common.config');
 module.exports = {
   ...commonConfig,
   displayName: 'nightly-tests',
+  testRunner: 'jest-circus/runner',
   setupFilesAfterEnv: ['jest-extended']
 };