+++ CHANGELOG.md
@@ -18,7 +18,7 @@
 
 ## New Functionality
 
--
+- [openapi-generator] Support custom API names and operation names by using OpenAPI extensions `x-sap-cloud-sdk-api-name` and `x-sap-cloud-sdk-operation-name`.
 
 ## Improvements
 
@@ -27,6 +27,9 @@
 ## Fixed Issues
 
 - [core] Fix building of entities using `fromJson` with navigation properties.
+- [odata-generator] The exit code of the node process is set to 1 in cases of a generator error.
+- [odata-generator] Fix the false-positive warning message, when functions/actions are not used by any function/action imports.
+
 
 # 1.38.0